ref: restructured package.json to improve readability [skip-ci]
Restructured the package.json to have the scripts at the top, below the basic properties like name and version. Then come the dependencies and dev-dependencies. At the end come other settings like eslint, husky and lint-staged.

This way the readability is improved as the most relevant information are positioned near the top.