chore: downgraded packages
The reason for this change is that Nextjs@15.0.3 was unable to render client components. It threw the same error that jest threw when converting trying to convert a server-component to a client component.

```
async/await is not yet supported in Client Components, only Server Components. This error is often caused by accidentally adding `'use client'` to a module that was originally written for the server
```

Downgraded the following packages:

- `next@15.0.3` -> `next@14.2.17`
-  `react@19.0.0-rc`-> `react@^18`
- `react-dom@19.0.0-rc` -> `react-dom@^18`