feat: added jest test for DesktopSideBar component
Having downgraded a few packages it is now possible to test a given component, such as the DesktopSideBar, that may be asynchronous.

Note that a component that is either async or not must not use another async component inside of it. Otherwise jest will be unable to test this component.