fix: resolved webpack devtool revert issue
As mentioned in #16 by downgrading from Next@15.0.3 to Next@14.2.17 overrides of the webpack config in terms of its devtools are reverted.

To partially overcome this issue, the `options.isServer` condition has been added. This allows the reporter to collect coverage for 2 out of 3 webpack runtimes. (https://nextjs.org/docs/14/app/api-reference/next-config-js/webpack)

On first glance the coverage-results seem to identical to the ones before the downgrade and webpack issue.