ref: added files argument to Codecov upload
Changed the code coverage upload to Codecov step. We now specify what specific file is to be uploaded (`lcov.info` file).  The reason for this change is that the logs indicate that multiple files were found. Even though the coverage is correctly displayed on the codecov-dashboard it does not post a comment in the corresponding PR.