fix: added mock of UserProfile component
The UserProfile component is now mocked for all tests, because it is asynchronous. Jest and the @testing-library/react packages do not support testing of asynchronous components within the components tree.

Also, Jest fails to parse the contents of the file that exports the function "getServerSession", which is why the getSession function is mocked as well.