ref: updated auto-merge prevention when auto-reviewed
When a PR is automatically reviewed by the auto-review workflow, then the 'manual-merge' label should be assigned to the PR so that it will not be merged automatically. Doing this in the auto-review workflow did not work, because the condition regarding the output-value was too unspecific. Now Mergify will assign the label accordingly.