fix: added condition to prevent loop of review-requests
Added a condition that prevents mergify from requesting a review from the same author on every change and therefore also from writing a comment every time the same reviewer is assigned. In other words the action is now only triggered when the person that is to be assigned to review has not been requested to review.