style: added light-mode styles to SideBarItem icon
When a user hovers over a SideBarItem the Icon will also be highlighted in the same color as the text and will have a stroke thickness of stroke-2. When the SideBarItem is selected it will also have the same color as the text and will have a stroke thickness of stroke-2.